+++ tigergrav/gravity.hpp
@@ -11,7 +11,7 @@
 
 struct multi_src {
 	multipole<float> m;
-	vect<float> x;
+	vect<double> x;
 	template<class A>
 	void serialize(A &&arc, unsigned) {
 		arc & m;
@@ -23,8 +23,8 @@ struct multi_src {
 
 struct multipole_info {
 	multipole<float> m;
-	vect<float> x;
-	float r;
+	vect<double> x;
+	double r;
 	bool has_active;
 	template<class A>
 	void serialize(A &&arc, unsigned) {
@@ -38,14 +38,14 @@ struct multipole_info {
 
 std::uint64_t gravity_PP_direct(std::vector<force> &g, const std::vector<vect<pos_type>> &x, std::vector<vect<pos_type>> &y);
 std::uint64_t gravity_PC_direct(std::vector<force> &g, const std::vector<vect<pos_type>> &x, std::vector<multi_src> &y);
-std::uint64_t gravity_CC_direct(expansion<double>&, const vect<float> &x, std::vector<multi_src> &y);
-std::uint64_t gravity_CP_direct(expansion<double> &L, const vect<float> &x, std::vector<vect<pos_type>> &y);
+std::uint64_t gravity_CC_direct(expansion<double>&, const vect<double> &x, std::vector<multi_src> &y);
+std::uint64_t gravity_CP_direct(expansion<double> &L, const vect<double> &x, std::vector<vect<pos_type>> &y);
 std::uint64_t gravity_PP_ewald(std::vector<force> &g, const std::vector<vect<pos_type>> &x, std::vector<vect<pos_type>> &y);
 std::uint64_t gravity_PC_ewald(std::vector<force> &g, const std::vector<vect<pos_type>> &x, std::vector<multi_src> &y);
-std::uint64_t gravity_CC_ewald(expansion<double>&, const vect<float> &x, std::vector<multi_src> &y);
-std::uint64_t gravity_CP_ewald(expansion<double> &L, const vect<float> &x, std::vector<vect<pos_type>> &y);
+std::uint64_t gravity_CC_ewald(expansion<double>&, const vect<double> &x, std::vector<multi_src> &y);
+std::uint64_t gravity_CP_ewald(expansion<double> &L, const vect<double> &x, std::vector<vect<pos_type>> &y);
 
 
-float ewald_near_separation(const vect<float> x);
-float ewald_far_separation(const vect<float> x, float r);
+double ewald_near_separation(const vect<double> x);
+double ewald_far_separation(const vect<double> x, double r);
 void init_ewald();
